make strings.py just about strings
Feels like this would be easier for new contributors to understand
where to look.
